DetailedCatogoryScreen edited

+++ lib/repository/api/detailed_catagory_screen/model/detailed_catogory_model.dart
@@ -39,7 +39,7 @@ class Datum {
     List<String>? nutrition;
     List<String>? technique;
     String? totalTime;
-    int? rating;
+    double? rating;
     String? author;
 
     Datum({
